# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "",
        "version": "1.0.0",
        "title": "API Kompetegram",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "kompetegram@upi.edu"
        }
    },
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "host": "localhost:3000",
    "basePath": "/api/v1",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/{kode_prodi}": {
            "get": {
                "tags": [
                    "Prodi"
                ],
                "operationId": "getProdi",
                "summary": "",
                "description": "Mendapatkan detail dari program studi secara spesifik",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "kode_prodi",
                        "in": "path",
                        "description": "Kode Prodi",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "200 response",
                        "examples": {
                        "application/json": {
                            "data": {
                                "kode": "J505",
                                "prodi": "Kewirausahaan",
                                "nama_fakultas": "Kampus UPI Tasikmalaya"
                            }
                        }
                        }
                    },
                    "404": {
                        "description": "404 response",
                        "examples": {
                            "application/json": {
                            "errors": {
                                "status": "404",
                                "title": "Tidak ditemukan",
                                "message": "Prodi dari fakultas bersangkutan tidak ditemukan"
                            }
                            }
                        }
                    }
                }
            }
        }
    }
}